+++ src/App.test.tsx
@@ -1,10 +1,6 @@
 import React from "react";
 import { mount } from "enzyme";
-import {
-  makeRouteProps,
-  mockJSDOM,
-  waitUntilFind,
-} from "./Helpers/TestHelpers";
+import { makeRouteProps, mockJSDOM } from "./Helpers/TestHelpers";
 import { MemoryRouter } from "react-router-dom";
 
 import App from "./App";
@@ -28,8 +24,11 @@ describe("App", () => {
     mockJSDOM();
   });
 
-  it("should have a link to bucket panel", async () => {
+  it("should have a link to dashboard panel", async () => {
     backendAPI.isAllowed = jest.fn().mockResolvedValue(true);
+    backendAPI.me = jest
+      .fn()
+      .mockResolvedValue({ permissions: { fullAccess: false } } as Token);
     routeProps.history.push = jest.fn();
 
     const app = mount(
@@ -38,10 +37,54 @@ describe("App", () => {
       </MemoryRouter>,
     );
 
-    const bucketItem = (await waitUntilFind(app, "#Buckets")).hostNodes().at(0);
-    // @ts-ignore
-    bucketItem.props().onClick();
+    await new Promise(process.nextTick);
+    app.update();
+
+    const bucketItem = app.find('li[data-menu-id$="dashboard"]');
+    bucketItem.simulate("click");
+    expect(routeProps.history.push).toBeCalledWith("/dashboard");
+  });
+
+  it("should have a link to bucket panel", async () => {
+    backendAPI.isAllowed = jest.fn().mockResolvedValue(true);
+    backendAPI.me = jest
+      .fn()
+      .mockResolvedValue({ permissions: { fullAccess: false } } as Token);
+    routeProps.history.push = jest.fn();
+
+    const app = mount(
+      <MemoryRouter>
+        <App {...routeProps} backendApi={backendAPI} />
+      </MemoryRouter>,
+    );
+
+    await new Promise(process.nextTick);
+    app.update();
+
+    const bucketItem = app.find('li[data-menu-id$="buckets"]');
+    bucketItem.simulate("click");
     expect(routeProps.history.push).toBeCalledWith("/buckets");
+  });
+
+  it("should have a link to replication panel", async () => {
+    backendAPI.isAllowed = jest.fn().mockResolvedValue(true);
+    backendAPI.me = jest
+      .fn()
+      .mockResolvedValue({ permissions: { fullAccess: true } } as Token);
+    routeProps.history.push = jest.fn();
+
+    const app = mount(
+      <MemoryRouter>
+        <App {...routeProps} backendApi={backendAPI} />
+      </MemoryRouter>,
+    );
+
+    await new Promise(process.nextTick);
+    app.update();
+
+    const securityItem = app.find('li[data-menu-id$="replications"]');
+    securityItem.simulate("click");
+    expect(routeProps.history.push).toBeCalledWith("/replications");
   });
 
   it("should have a link to security panel", async () => {
@@ -57,11 +100,11 @@ describe("App", () => {
       </MemoryRouter>,
     );
 
-    const bucketItem = (await waitUntilFind(app, "#Security"))
-      .hostNodes()
-      .at(0);
-    // @ts-ignore
-    bucketItem.props().onClick();
+    await new Promise(process.nextTick);
+    app.update();
+
+    const securityItem = app.find('li[data-menu-id$="security"]');
+    securityItem.simulate("click");
     expect(routeProps.history.push).toBeCalledWith("/tokens");
   });
 
@@ -77,7 +120,45 @@ describe("App", () => {
       </MemoryRouter>,
     );
 
-    const bucketItem = await waitUntilFind(app, "#Security");
-    expect(bucketItem).toBeUndefined();
+    await new Promise(process.nextTick);
+    app.update();
+
+    expect(app.find('li[data-menu-id$="security"]').exists()).toBeFalsy();
+  });
+
+  it("should hide replications panel", async () => {
+    backendAPI.isAllowed = jest.fn().mockResolvedValue(true);
+    backendAPI.me = jest
+      .fn()
+      .mockResolvedValue({ permissions: { fullAccess: false } } as Token);
+
+    const app = mount(
+      <MemoryRouter>
+        <App {...routeProps} backendApi={backendAPI} />
+      </MemoryRouter>,
+    );
+
+    await new Promise(process.nextTick);
+    app.update();
+
+    expect(app.find('li[data-menu-id$="replications"]').exists()).toBeFalsy();
+  });
+
+  it("should show help menu", async () => {
+    backendAPI.isAllowed = jest.fn().mockResolvedValue(true);
+    backendAPI.me = jest
+      .fn()
+      .mockResolvedValue({ permissions: { fullAccess: false } } as Token);
+
+    const app = mount(
+      <MemoryRouter>
+        <App {...routeProps} backendApi={backendAPI} />
+      </MemoryRouter>,
+    );
+
+    await new Promise(process.nextTick);
+    app.update();
+
+    expect(app.find(".help-menu-item").exists()).toBeTruthy();
   });
 });