+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - RS-582: fix remove modal to display error messages, [PR-75](https://github.com/reductstore/web-console/pull/75)
+- RS-600: fix CI workflow and corrected a prop name in EntryCard component, [PR-77](https://github.com/reductstore/web-console/pull/77)
 
 ### Added: