+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - RS-580: browse records in the console, [PR-76](https://github.com/reductstore/web-console/pull/76)
+- RS-588: add useful links in the side menu, [PR-79](https://github.com/reductstore/web-console/pull/79)
 
 ### [1.8.1] - 2024-12-16