+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### [1.9.0] - 2025-02-25
+
 ### Fixed
 
 - RS-582: fix remove modal to display error messages, [PR-75](https://github.com/reductstore/web-console/pull/75)
@@ -201,7 +203,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial Release
 
 [Unreleased]: https://github.com/reduct-storage/web-console/compare/v1.8.1...HEAD
-[1.8.0]: https://github.com/reduct-storage/web-console/compare/v1.8.1...v1.8.0
+[1.9.0]: https://github.com/reduct-storage/web-console/compare/v1.9.0...v1.8.1
+[1.8.1]: https://github.com/reduct-storage/web-console/compare/v1.8.1...v1.8.0
 [1.8.0]: https://github.com/reduct-storage/web-console/compare/v1.8.0...v1.7.0
 [1.7.0]: https://github.com/reduct-storage/web-console/compare/v1.7.0...v1.6.1
 [1.6.1]: https://github.com/reduct-storage/web-console/compare/v1.6.1...v1.6.0